refactor: ATLAS-838: Move ResultSet to more appropriate namespace.
