+++ Atlas.Client.Models/Search/Results/Matching/ResultSet/RepeatMatchingAlgorithmResultSet.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Atlas.Client.Models.Search.Results.ResultSet;
 
 namespace Atlas.Client.Models.Search.Results.Matching.ResultSet
 {
