+++ Atlas.Functions/Services/MatchPredictionInputBuilder.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using Atlas.Client.Models.Search.Requests;
-using Atlas.Client.Models.Search.Results;
 using Atlas.Client.Models.Search.Results.Matching;
+using Atlas.Client.Models.Search.Results.ResultSet;
 using Atlas.Common.ApplicationInsights;
 using Atlas.Common.ApplicationInsights.Timing;
 using Atlas.Common.GeneticData;
