+++ Atlas.MatchingAlgorithm.Test.Performance/SearchTimingService.cs
@@ -6,8 +6,8 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Atlas.Client.Models.Search.Requests;
-using Atlas.Client.Models.Search.Results;
 using Atlas.Client.Models.Search.Results.Matching;
+using Atlas.Client.Models.Search.Results.ResultSet;
 using Atlas.MatchingAlgorithm.Test.Performance.Models;
 using Newtonsoft.Json;
