+++ Atlas.MatchingAlgorithm.Test.Validation/TestData/Models/SearchAlgorithmApiResult.cs
@@ -1,6 +1,6 @@
 ﻿using System.Net;
-using Atlas.Client.Models.Search.Results;
 using Atlas.Client.Models.Search.Results.Matching;
+using Atlas.Client.Models.Search.Results.ResultSet;
 
 namespace Atlas.MatchingAlgorithm.Test.Validation.TestData.Models
 {
