+++ Atlas.MatchingAlgorithm/Clients/AzureStorage/ResultsBlobStorageClient.cs
@@ -2,8 +2,8 @@ using Atlas.Common.AzureStorage.Blob;
 using Atlas.MatchingAlgorithm.Settings.Azure;
 using Newtonsoft.Json;
 using System.Threading.Tasks;
-using Atlas.Client.Models.Search.Results;
 using Atlas.Client.Models.Search.Results.Matching;
+using Atlas.Client.Models.Search.Results.ResultSet;
 using Atlas.MatchingAlgorithm.ApplicationInsights.ContextAwareLogging;
 
 namespace Atlas.MatchingAlgorithm.Clients.AzureStorage
