+++ Atlas.RepeatSearch/Clients/AzureStorage/RepeatSearchResultsBlobStorageClient.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
-using Atlas.Client.Models.Search.Results;
 using Atlas.Client.Models.Search.Results.Matching;
+using Atlas.Client.Models.Search.Results.ResultSet;
 using Atlas.Common.AzureStorage.Blob;
 using Atlas.MatchingAlgorithm.ApplicationInsights.ContextAwareLogging;
 using Atlas.RepeatSearch.Settings.Azure;
